+++ Changelog.md
@@ -1,4 +1,5 @@
 # Version 0.9.0
+- Feature: Support parameter info 'View > Parameter Info' for instructions, types and tags
 - Fix wrong indentation after pressing enter in the code section
 - Don't suggest instructions at top level of a file
 - Don't suggest instructions inside of comments